refactor(sync): extract device-registration module
Move PLATFORM_MAP, registerDevice, and persistKeysAndRegisterDevice
from apps/desktop/src/main/ipc/sync-handlers.ts into a new
apps/desktop/src/main/sync/device-registration.ts. This is a
prerequisite for the Phase 2 sync-handlers split: these helpers are
needed by multiple child handler files (auth-oauth, auth-device) and
already have external consumers (sync/linking-service.ts,
main/test-hooks.ts).

- sync-handlers.ts shrinks 1318 → 1157 lines
- No behavior change; imports redirected in linking-service.ts and
  test-hooks.ts
- Part of Phase 2 (tech-debt-remediation.md): prep commit before
  splitting handlers
